style (NEW: '___' prefix and '___' suffix for "raw" macros)
This is *not* a raw C macro:

    #define MY_MACRO(...) \
        do {              \
          /* BODY */      \
        } while (false)

It's the classic way of writing a C macro so it can be safely thought of as just
any old regular function. But a raw C macro *cannot* be thought of as "just any
old function", because that's just not what it is. Here is an example which
illustrates this:

    #define JOVE_SCOPED_BUFF(var, len)                                         \
      const jove_buffer_t __##var##_buff _CLEANUP(_jove_free_buffer) =         \
          _jove_alloc_buffer(len);                                             \
      var = (char *)__##var##_buff.ptr

This C macro does something that may be unexpected: it defines a variable in the
current scope. Furthermore, __attribute__((cleanup)) defines a function which is
automatically called by the compiler when the program goes *out of* the scope.

This RAII-style C code (which is in many respects C code wishing it was C++) is
going to be used a lot in the jove runtime. So I came up with the following
naming convention: for any *raw* macro, you must name it as starting with '___'
and ending with '___'. Not one underscore, not two [1], but three- before *and*
after the name.

[1] Two preceeding underscores ('__') are already being used as a convention for
global *variables* in the jove runtime (one preceeding underscore ('_') is the
convention for *functions*). The idea behind this is two avoid name conflicts
with symbols from existing programs.