loop: get helper code straight from building QEMU emulator with -plugin-opt=emit-llvm
Until now, the bitcode for helpers (i.e. the QEMU runtime) has been gotten by
using carbon-copy to extract each one into its own .c file, and build that with
clang. Instead, we'll build the QEMU emulator with '-Wl,-plugin-opt=emit-llvm',
which gives us a bitcode file which contains all the helpers.

llknife is used to cut up this bitcode file so that we have an LLVM bitcode for
each helper, as before.

This commit addresses a possible criticism that the correctness of the
recompilation ultimately depends on the soundness of what carbon-copy does, and
it is not easy to prove carbon-copy is producing perfect output (for the inputs
we give it).