use boost stacktrace for capturing stacktraces at throw
the boost developers are on the right track. the C++ programmer *needs* a
stacktrace upon catching an exception.

obviously, we like destructors very much. if some code takes exclusive
lock on the Binaries, for example, we need that lock to be released if
that code somehow errors out.

we define our own assert() now, in jove/assert.h. instead of raising
a SIGABRT, it throws an exception.