style (alignment)
Why does alignment matter? It allows us to more easily understand the code
because of cognitive reasons- it's "faster" to see "what is what".

Imagine you're accessing a bunch of members from a struct (the _same_ struct).
If you align the '.' (or '->') so that they line up (effectively creating a
vertical line), the reader can more quickly process the visual information. It's
akin to a waiter holding two items on the same platter versus holding two
platters (for each item, at different heights).