patches/linux.diff: take the fast-path guardrails off the syscall-exit path and replace them with a sledgehammer.
Ideally, we don't want to have to modify the kernel in any way for `jove ipt`.
But, we are already doing it now so we have the bpf_get_fd_path() available to
BPF programs [1].

A consequence of this change is that we are now guaranteed to see a
raw_syscalls:sys_exit tracepoint after every successful execve(2).

It seems that this change also (similarly) improves maintinence of the
pids_unfiltered bpf map in jove_augmented_raw_syscalls.bpf.c.

[1] As a reminder, this bpf helper is necessary because the WINE server sends
the file descriptors for DSOs over a unix socket, and, unless we saw the call
to open(2), we simply cannot know after-the-fact what was mmap'd into memory.